[NFC] Split the basic target-layout logic into its own file
